requirements.txt
google-api-python-client==2.114.0
google-auth-oauthlib==1.2.0
pyyaml==6.0.1
